+++ packages/liveblocks-emails/src/liveblocks-text-editor.tsx
@@ -12,7 +12,6 @@ import type {
   ResolveUsersArgs,
 } from "@liveblocks/core";
 import { html, htmlSafe } from "@liveblocks/core";
-import type { ComponentType, ReactNode } from "react";
 
 import type {
   LexicalMentionNodeWithContext,
@@ -306,144 +305,6 @@ export const resolveUsersInLiveblocksTextEditorNodes = async <
   }
   return resolvedUsers;
 };
-
-export type TextEditorContainerComponentProps = {
-  /**
-   * The nodes of the text editor
-   */
-  children: ReactNode;
-};
-
-export type TextEditorMentionComponentProps<U extends BaseUserMeta = DU> = {
-  /**
-   * The mention element.
-   */
-  element: LiveblocksTextEditorMentionNode;
-  /**
-   * The mention's user info, if the `resolvedUsers` option was provided.
-   */
-  user?: U["info"];
-};
-
-export type TextEditorTextComponentProps = {
-  /**
-   * The text element.
-   */
-  element: LiveblocksTextEditorTextNode;
-};
-
-export type ConvertTextEditorNodesAsReactComponents<
-  U extends BaseUserMeta = DU,
-> = {
-  /**
-   *
-   * The component used to act as a container to wrap text editor nodes,
-   */
-  Container: ComponentType<TextEditorContainerComponentProps>;
-
-  /**
-   * The component used to display mentions.
-   */
-  Mention: ComponentType<TextEditorMentionComponentProps<U>>;
-
-  /**
-   * The component used to display text nodes.
-   */
-  Text: ComponentType<TextEditorTextComponentProps>;
-};
-
-const baseComponents: ConvertTextEditorNodesAsReactComponents<BaseUserMeta> = {
-  Container: ({ children }) => <div>{children}</div>,
-  Mention: ({ element, user }) => (
-    <span data-mention>
-      {MENTION_CHARACTER}
-      {user?.name ?? element.userId}
-    </span>
-  ),
-  Text: ({ element }) => {
-    // Note: construction following the schema 👇
-    // <code><s><em><strong>{element.text}</strong></s></em></code>
-    let children: ReactNode = element.text;
-
-    if (element.bold) {
-      children = <strong>{children}</strong>;
-    }
-
-    if (element.italic) {
-      children = <em>{children}</em>;
-    }
-
-    if (element.strikethrough) {
-      children = <s>{children}</s>;
-    }
-
-    if (element.code) {
-      children = <code>{children}</code>;
-    }
-
-    return <span>{children}</span>;
-  },
-};
-
-export type ConvertTextEditorNodesAsReactOptions<U extends BaseUserMeta = DU> =
-  {
-    /**
-     * The components used to customize the resulting React nodes. Each components has
-     * priority over the base components inherited.
-     */
-    components?: Partial<ConvertTextEditorNodesAsReactComponents<U>>;
-    /**
-     * A function that returns user info from user IDs.
-     */
-    resolveUsers?: (
-      args: ResolveUsersArgs
-    ) => Awaitable<(U["info"] | undefined)[] | undefined>;
-  };
-
-/**
- * @internal
- *
- * Convert a set of Liveblocks Editor nodes into React elements
- */
-export async function convertTextEditorNodesAsReact(
-  nodes: LiveblocksTextEditorNode[],
-  options?: ConvertTextEditorNodesAsReactOptions<BaseUserMeta>
-): Promise<ReactNode> {
-  const Components = {
-    ...baseComponents,
-    ...options?.components,
-  };
-  const resolvedUsers = await resolveUsersInLiveblocksTextEditorNodes(
-    nodes,
-    options?.resolveUsers
-  );
-
-  const children = nodes.map((node, index) => {
-    switch (node.type) {
-      case "mention":
-        return (
-          <Components.Mention
-            key={`lb-text-editor-mention-${index}-${node.userId}`}
-            element={node}
-            user={resolvedUsers.get(node.userId)}
-          />
-        );
-      case "text":
-        return (
-          <Components.Text
-            key={`lb-text-editor-text-${index}`}
-            element={node}
-          />
-        );
-    }
-  });
-
-  return (
-    <Components.Container key="lb-text-editor-container">
-      {children}
-    </Components.Container>
-  );
-}
 
 export type ConvertTextEditorNodesAsHtmlStyles = {
   /**
